pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cea-backend"
version = "0.1.0"
description = "Climate Economy Assistant Backend"
authors = [{name = "George Nekwaya"}]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.26.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.2",
    "psycopg2-binary>=2.9.7",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langsmith>=0.1.0",
    "langgraph-runtime-inmem>=0.2.0,<0.3",
    "tiktoken>=0.5.1",
    "numpy>=1.24.3",
    "faiss-cpu>=1.7.4",
    "openai>=1.3.7",
    "anthropic>=0.7.7",
    "instructor>=0.4.5",
    "pandas>=2.1.3",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "redis>=5.0.1",
    "celery>=5.3.4",
    "cryptography>=42.0.0,<45.0",
    "authlib>=1.2.1",
    "PyJWT>=2.8.0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "httpx>=0.25.2",
    "aiofiles>=23.2.1",
    "asyncio-mqtt>=0.13.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "gunicorn>=21.2.0",
    "spacy>=3.7.2",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl",
    "textblob>=0.17.1",
    "scikit-learn>=1.3.2",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.23.0",
    "geopy>=2.4.0",
    "geocoder>=1.38.1",
    "jsonschema-rs>=0.20.0,<0.30"
]

[tool.setuptools.packages.find]
where = ["backend"]

[tool.setuptools.package-dir]
"" = "backend"

[project.scripts]
climate-agent = "api.workflows.climate_workflow:main"
resume-agent = "api.workflows.resume_workflow:main"
career-agent = "api.workflows.career_workflow:main"
climate-supervisor-agent = "api.workflows.climate_supervisor_workflow:main"